.vscode/requirements.txt
PyYAML
invoke
psutil
requests
